✨ feat: Add redux-thunk to project, implement getUsersThunkCreator
This commit adds the `redux-thunk` library to the project and implements a new action creator called `getUsersThunkCreator`. This action creator is used to fetch users data from the server and dispatch actions accordingly. The `setIsFetchingActionCreator` action is dispatched when the request is initiated, and `setUsersActionCreator`, `setTotalUsersCountActionCreator`, and `setIsFetchingActionCreator` are dispatched with the received data when the request is successfully completed.