✨ feat: Add edit mode state and functions to ProfileStatus component
The commit adds new functionality to the ProfileStatus component by importing useState from React and using it to manage the editMode state. It also defines two functions, activeEditMode and disableEditMode, that modify the editMode state variable. When the editMode state is true, the component renders an input field to allow users to edit their status. Otherwise, it displays a span element containing the current status text that can be clicked to activate the edit mode.