✨ feat: add functionality to request login data
This commit adds the ability to make a request for login data by creating `setLoginDataRequest` in `/api/api.js`. It also updates `LoginComponent` in `/component/Login/Login.jsx` to call `requestLoginData` along with `setLoginData`. Finally, it adds `setLoginDataThunkCreator` to `/redux/formReducer.js`, which uses `setLoginDataRequest` to get the login data and update the state accordingly.