✨ feat: Add logout functionality to Header
This commit adds the ability for users to log out of the application from the Header component. The `logoutRequest()` function has been added to `api.js`, and a new action creator `logoutThunkCreator()` has been added to `authReducer.js`. The `Header` component now receives the `logoutUser` prop, which is passed down to the Logout button, allowing users to log out of the application when clicked.