✨ feat: Add error message to login form
This commit adds an error message to the login form when there is an authentication error. The `Login` component now passes a `errorMessage` prop to its child component, `LoginForm`. The `LoginForm` component uses the `useForm` hook's `setError` method to set a custom error message if `props.errorMessage` has a value. The `authReducer` now defines a new action type `SET_SOME_ERRORS` and updates the state with the error message passed from the server in case of unsuccessful login. The `setSomeErrorsActionCreator` function creates an action object with the type `SET_SOME_ERRORS` and an `errorMessages` payload.