✨ feat: Add AppContainer and setInitialLoadActionCreator
This commit adds a new file "AppContainer.jsx" that replaces the "App" component in the "index.js" file. The new "AppContainer" is a higher-order component that connects to Redux, maps the state to props, and wraps the "App" component.

Additionally, this commit also includes a new file "appReducer.js" that contains the action type and creator for setting the initial load. This reducer is added to the combineReducers function in "reduxStore.js".