✨ feat: Add loader component to App component
Add a Loader component to the App component, displayed when the initialLoad prop is false.

refactor: Remove unnecessary code from HeaderContainer

Remove the setUserAuthData function and its useEffect call, as it is not being used.

feat: Add setInitialLoadActionCreator to authReducer

Import and use the setInitialLoadActionCreator function from appReducer to set the initialLoad state to true when logging out.